refactor all the views
# Added in separate files
- CustomPopUpView
and other cleanups
