+++ DragandDrop-main/DragandDrop/Model/TaskItem.swift
@@ -3,10 +3,23 @@ import Algorithms
 import UniformTypeIdentifiers
 
 struct TaskItem: Codable, Hashable, Transferable {
+    
     let id: UUID
     let title: String
     let owner: String
     let note: String
+    
+    init(
+        id: UUID = UUID(),
+        title: String,
+        owner: String = "tanay",
+        note: String = ""
+    ) {
+        self.id = id
+        self.title = title
+        self.owner = owner
+        self.note = note
+    }
     
     static var transferRepresentation: some TransferRepresentation {
         CodableRepresentation(contentType: .taskItem)
